requirements.txt
celery==5.2.7
django-admin-rangefilter==0.10.0
django-admin-relation-links==0.2.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-json-widget==1.1.1
django-jsonform==2.10.1
django-redis==5.2.0
django-timezone-field==5.0
django-tinymce==3.3.0
Django==3.2.13
djangorestframework==3.13.1
grpcio-tools==1.47.0
grpcio==1.47.0
psycopg2==2.9.3
Pygments==2.12.0
redis==4.3.4
requests==2.28.1
tzlocal==4.2